Update debian package files.
Improved build_deb_package.sh script
Added homepage to control
minor changes to cyberfox.desktop

+++ _Build/_Linux/build_deb_package.sh
@@ -1,22 +1,45 @@
 #!/bin/bash
 # Built from template by hawkeye116477
 # Full repo https://github.com/hawkeye116477/cyberfox-deb
-
+# Script Version: 1.0.2
 # Set current directory to script directory.
 Dir=$(cd "$(dirname "$0")" && pwd)
 cd $Dir
 
+# Init vars
+VERSION=""
+CONTROLTEMPLATE=""
+
 # Get package version including beta version if beta.
-VERSION=$(<../../browser/config/version_display.txt)
+if [ -f "../../browser/config/version_display.txt" ]; then
+    VERSION=$(<../../browser/config/version_display.txt)
+else
+    echo "Unable to get current build version!"
+    exit 1    
+fi
 
 # Generate template
-CONTROLTEMPLATE=$Dir/deb/DEBIAN/control
-cp _Templates/control deb/DEBIAN
+if [ -f "$Dir/deb/DEBIAN/control" ]; then
+    CONTROLTEMPLATE=$Dir/deb/DEBIAN/control
+    cp _Templates/control deb/DEBIAN
+else
+    echo "Unable to location control template!"
+    exit 1    
+fi
+
+# Generate cyberfox folder if not already created
+if [ ! -d "$Dir/deb/usr/lib/Cyberfox" ]; then
+    mkdir $Dir/deb/usr/lib/Cyberfox
+fi
 
 # Copy latest build
-mkdir $Dir/deb/usr/lib/Cyberfox
-cp -R ../../../obj64/dist/Cyberfox/* $Dir/deb/usr/lib/Cyberfox
-cp _Templates/Cyberfox.sh $Dir/deb/usr/lib/Cyberfox
+if [ -d "../../../obj64/dist/Cyberfox" ]; then
+    cp -R ../../../obj64/dist/Cyberfox/* $Dir/deb/usr/lib/Cyberfox
+    cp _Templates/cyberfox.sh $Dir/deb/usr/lib/Cyberfox
+else
+    echo "Unable to Cyberfox package files, Please check the build was created and packaged successfully!"
+    exit 1     
+fi
 
 # Apply information to template
 cd $Dir/deb
@@ -24,6 +47,9 @@ if grep -q -E "__VERSION__|__SIZE__" "$CONTROLTEMPLATE" ; then
     sed -i "s|__VERSION__|$VERSION|" $CONTROLTEMPLATE   
     SIZE=$(du -s)
     sed -i "s|__SIZE__|${SIZE::-1}|" $CONTROLTEMPLATE
+else
+    echo "An error occured when trying to generate $CONTROLTEMPLATE information!"
+    exit 1  
 fi
 
 # Generate file hashes
@@ -31,11 +57,23 @@ find . -type f ! -regex '.*.hg.*' ! -regex '.*?debian-binary.*' ! -regex '.*DEBI
 cp -r DEBIAN/md5sums ../../../../obj64/dist
 cd $Dir
 
-
 # Build debian package
-dpkg -b $Dir/deb ../../../obj64/dist/Cyberfox-$VERSION.en-US.linux-x86_64
+dpkg -b $Dir/deb ../../../obj64/dist/Cyberfox-$VERSION.en-US.linux-x86_64.deb
+
+# Notify of package complete
+if [ -f "../../../obj64/dist/Cyberfox-$VERSION.en-US.linux-x86_64.deb" ]; then
+    notify-send "Debian package Complete"
+fi
 
 # Clean up
-rm -rf $Dir/deb/usr/lib/Cyberfox
-rm -f $Dir/deb/DEBIAN/md5sums
+if [ -d "$Dir/deb/usr/lib/Cyberfox" ]; then
+    echo "Clean: $Dir/deb/usr/lib/Cyberfox"
+    rm -rf $Dir/deb/usr/lib/Cyberfox
+fi
+if [ -f "$Dir/deb/DEBIAN/md5sums" ]; then
+    echo "Clean: $Dir/deb/DEBIAN/md5sums"
+    rm -f $Dir/deb/DEBIAN/md5sums
+fi
+
+# Add fresh template
 cp $Dir/_Templates/control deb/DEBIAN
